[FP-73] Refactor: 타입 경로 변경된 사한 적용
